requirements.txt
Flask==2.3.4
pillow
pytesseract
werkzeug